parser: add a binary operation system
This system dynamically checks the types of the operands and the token
type of the operator and then uses a lookup table to perform the
operation.

The resulting code is much simpler, organised and no longer a mess on
if-else-switch statement with duplicated error handling.

Further cleaning is needed, along with missing error messages.

This system would allow us to easily add new operators without chaning
too much code, and maybe in future allows users to dynamically add new
operators without forking and changing the code.