Data: Added new methods ``is()`` and ``get()``
The is(t) method is used to check if t is the current active tag of the
value of the data.

The get(t) method is used to get the actual value hold by tag 't'. No
special checking is performed to ensure its the currently active tag.

Additionally, t should be comptime known for get() method. While this is
not true for is() method, but its meant to be used like that anyways.