main: allow function call without parens when multiple calls are not being nested
This means ``index [10, 20, 30] 1`` is allowed but ``* + 10 20 30`` is not
allowed since it is confusing. So the correct way to write the second
expression is ``* (+ 10 20) 30``

For regular function, the old syntax still works as expected: ``(+ 10
20)``

This change permanently rejects the idea of varargs in function calls,
since they won't be possible to implement.