Data+main: initial support towards first class native functions
This commit adds native functions as a type to Data and makes the
necessary changes elsewhere. Functionality wise, functions still cannot
be assigned to variables as any mention of a function is automatically
considered a function call. Hence, additional syntax/work is needed

There are still a bunch of other minor TODOs left regarding native
functions in Data